requirements.txt
Flask==2.0.1
Flask-Humanize==0.3.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
gunicorn==20.1.0
Authlib==0.15.5
requests==2.26.0
Flask-RESTful==0.3.9
python-dotenv==0.19.2
Flask==2.0.2
Flask-WTF==1.0.0
arrow==1.2.1
marshmallow==3.14.1
